Safari-609.1.11.1/LayoutTests/http/tests/notifications/events-in-dictionary-expected.txt

This test ensures event handlers defined inside the options dictionary are ignored.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Notification.permission is "granted"
PASS onshow event handler defined after Notification construction was called
PASS successfullyParsed is true

TEST COMPLETE
